+++ neural_lam/models/step_predictors/graph/hi_lam.py
@@ -37,6 +37,7 @@ class HiLAM(BaseHiGraphModel):
         m2g_gnn_type: str = "InteractionNet",
         mesh_up_gnn_type: str = "InteractionNet",
         mesh_down_gnn_type: str = "InteractionNet",
+        use_heterodata: bool = False,
     ):
         """
         Initialize the HiLAM model.
@@ -82,6 +83,7 @@ class HiLAM(BaseHiGraphModel):
             m2g_gnn_type=m2g_gnn_type,
             mesh_up_gnn_type=mesh_up_gnn_type,
             mesh_down_gnn_type=mesh_down_gnn_type,
+            use_heterodata=use_heterodata,
         )
 
         # Make down GNNs, both for down edges and same level